Rename optimize to export + fix tests
