+++ anomalib/utils/callbacks/__init__.py
@@ -14,6 +14,8 @@ from jsonargparse.namespace import Namespace
 from omegaconf import DictConfig, ListConfig, OmegaConf
 from pytorch_lightning.callbacks import Callback, ModelCheckpoint
 
+from anomalib.deploy import ExportMode
+
 from .cdf_normalization import CdfNormalizationCallback
 from .graph import GraphLogger
 from .metrics_configuration import MetricsConfigurationCallback
@@ -134,7 +136,7 @@ def get_callbacks(config: Union[ListConfig, DictConfig]) -> List[Callback]:
                     input_size=config.model.input_size,
                     dirpath=config.project.path,
                     filename="model",
-                    export_mode=config.optimization.export_mode,
+                    export_mode=ExportMode(config.optimization.export_mode),
                 )
             )
         else:
