+++ tests/pre_merge/deploy/test_inferencer.py
@@ -15,7 +15,7 @@ from pytorch_lightning import Trainer
 from anomalib.config import get_configurable_parameters
 from anomalib.data import get_datamodule
 from anomalib.deploy import OpenVINOInferencer, TorchInferencer, export
-from anomalib.deploy.optimize import ExportMode
+from anomalib.deploy.export import ExportMode
 from anomalib.models import get_model
 from anomalib.utils.callbacks import get_callbacks
 from tests.helpers.dataset import TestDataset, get_dataset_path
