+++ tools/benchmarking/benchmark.py
@@ -26,7 +26,7 @@ from utils import upload_to_comet, upload_to_wandb, write_metrics
 from anomalib.config import get_configurable_parameters, update_input_size_config
 from anomalib.data import get_datamodule
 from anomalib.deploy import export
-from anomalib.deploy.optimize import ExportMode
+from anomalib.deploy.export import ExportMode
 from anomalib.models import get_model
 from anomalib.utils.loggers import configure_logger
 from anomalib.utils.sweep import (
